mv execute_pipeline to Metagenomic
